refactor MediaManagerListener
